T #50 New params for GPT3

+++ Huppy/Huppy.Core/Dto/GPTDto.cs
@@ -29,9 +29,9 @@ namespace Huppy.Core.Dto
         public string[]? Stop { get; set; }
 
         [JsonPropertyName("frequency_penalty")]
-        public int FrequencyPenalty { get; set; }
+        public double FrequencyPenalty { get; set; }
 
         [JsonPropertyName("presence_penalty")]
-        public int PresencePenalty { get; set; }
+        public double PresencePenalty { get; set; }
     }
 }
